transport: stabilize local fallback protocol handling
