+++ src/api/eventsource/tests.rs
@@ -1,0 +1,10 @@
+use super::*;
+use crate::runtime::RuntimeEvent;
+use axum::Router;
+use axum::routing::get;
+use futures::stream;
+use std::sync::Arc;
+use std::sync::atomic::{AtomicUsize, Ordering};
+
+mod local_retry;
+mod stream_lifecycle;
